+++ src/converter/converters.js
@@ -5,7 +5,6 @@ import {BannerPatternConverter} from "./BannerPatternConverter";
 import {BarConverter} from "./BarConverter";
 import {BedConverter} from "./BedConverter";
 import {BeeConverter} from "./BeeConverter";
-import {ButtonConverter} from "./ButtonConverter";
 import {ChestFrontConverter} from "./ChestFrontConverter";
 import {ChestLeftRightDoubleConverter} from "./ChestLeftRightDoubleConverter";
 import {ChestNormalConverter} from "./ChestNormalConverter";
@@ -14,14 +13,19 @@ import {ColorizeOverlayConverter} from "./ColorizeOverlayConverter";
 import {CopyConverter} from "./CopyConverter";
 import {DeleteConverter} from "./DeleteConverter";
 import {DespriteConverter} from "./DespriteConverter";
+import {DespriteExperimentalConverter} from "./DespriteExperimentalConverter";
 import {DestroyStageConverter} from "./DestroyStageConverter";
+import {DialogConverter} from "./DialogConverter";
 import {DrownedConverter} from "./DrownedConverter";
+import {EnchantedItemGlintConverter} from "./EnchantedItemGlintConverter";
 import {FishHookConverter} from "./FishingConverter";
 import {FireworksConverter} from "./FireworksConverter";
 import {FoxConverter} from "./FoxConverter";
 import {HorseConverter} from "./HorseConverter";
+import {IconsConverter} from "./IconsConverter";
 import {MapIconsConverter} from "./MapIconsConverter";
 import {MetadataConverter} from "./MetadataConverter";
+import {NineSliceConverter} from "./NineSliceConverter";
 import {OpaqueConverter} from "./OpaqueConverter";
 import {OverlayToTranslateConverter} from "./OverlayToTranslateConverter";
 import {Particles1_13Converter} from "./Particles1_13Converter";
@@ -57,6 +61,7 @@ const converters = [
     ...FishHookConverter.getDefaultConverters(),
     ...FoxConverter.getDefaultConverters(),
     ...HorseConverter.getDefaultConverters(),
+    ...IconsConverter.getDefaultConverters(),
     ...MapIconsConverter.getDefaultConverters(),
     ...PistonArmConverter.getDefaultConverters(),
     ...RedstoneDustConverter.getDefaultConverters(),
@@ -69,8 +74,10 @@ const converters = [
     ...BeeConverter.getDefaultConverters(),
     ...TitleConverter.getDefaultConverters(),
     ...DespriteConverter.getDefaultConverters(),
+    ...DespriteExperimentalConverter.getDefaultConverters(),
     ...BarConverter.getDefaultConverters(),
-    ...ButtonConverter.getDefaultConverters(),
+    ...NineSliceConverter.getDefaultConverters(),
+    ...DialogConverter.getDefaultConverters(),
     ...OverlayToTranslateConverter.getDefaultConverters(),
     ...ColorizeOverlayConverter.getDefaultConverters(),
     ...PlaceholderConverter.getDefaultConverters(),
@@ -79,6 +86,7 @@ const converters = [
     ...Particles1_13Converter.getDefaultConverters(),
     ...SpriteConverter.getDefaultConverters(),
     ...DestroyStageConverter.getDefaultConverters(),
+    ...EnchantedItemGlintConverter.getDefaultConverters(),
     ...PngToTgaConverter.getDefaultConverters(),
     ...CopyConverter.getDefaultConverters(),
     ...DeleteConverter.getDefaultConverters()