+++ src/index.js
@@ -2,6 +2,7 @@ import {addAdditionalConverters, getConverters} from "./converter";
 import {AbstractInput} from "./input";
 import {AbstractLog} from "./log";
 import {AbstractOutput} from "./output";
+import {Options} from "./Options";
 
 /**
  * Class ConvertJavaTextureToBedrock
@@ -13,8 +14,9 @@ class ConvertJavaTextureToBedrock {
      * @param {AbstractInput} input
      * @param {AbstractOutput} output
      * @param {AbstractLog} log
+     * @param {Options} options
      */
-    constructor(input, output, log) {
+    constructor(input, output, log, options = {}) {
         /**
          * @type {AbstractInput}
          *
@@ -33,32 +35,40 @@ class ConvertJavaTextureToBedrock {
          * @protected
          */
         this.log = log;
+        /**
+         * @type {Options}
+         *
+         * @protected
+         */
+        this.options = {
+            ...new Options(),
+            ...options
+        };
     }
 
     /**
      * @returns {Promise<*>}
      */
     async convert() {
-        try {
-            await this.output._init(this.input, this.log);
-
-            for await (const entry of this.input.getEntries()) {
-                await entry._init(this.log);
-
-                await this.output.applyInputEntry(entry);
-            }
-
-            for await (const converter of getConverters()) {
-                await converter._init(this.input, this.output, this.log);
-
-                await addAdditionalConverters(...await converter.convert());
-            }
-
-            return await this.output.generate();
-        } catch (err) {
-            this.log.error(err.message);
-            throw err;
+        if (this.options.experimental) {
+            this.log.warn(`EXPERIMENTAL FEATURES ENABLED!`)
         }
+
+        await this.output._init(this.input, this.log, this.options);
+
+        for await (const entry of this.input.getEntries()) {
+            await entry._init(this.log, this.options);
+
+            await this.output.applyInputEntry(entry);
+        }
+
+        for await (const converter of getConverters()) {
+            await converter._init(this.input, this.output, this.log, this.options);
+
+            await addAdditionalConverters(...await converter.convert());
+        }
+
+        return await this.output.generate();
     }
 }